+++ src/main/java/com/pictalk/user/domain/User.java
@@ -1,0 +1,42 @@
+package com.pictalk.user.domain;
+
+import com.pictalk.message.domain.Sender;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Table(name = "user")
+@Getter
+@Setter
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(nullable = false)
+    private Long userId;
+
+    @OneToMany(mappedBy = "user")
+    private List<Sender> senders = new ArrayList<>();
+
+    @Column(nullable = false)
+    private String username;
+
+    @Column(nullable = false)
+    private String password;
+
+    @Column(nullable = false)
+    private LocalDateTime createdAt;
+
+    @PrePersist
+    protected void onCreate() {
+        createdAt = LocalDateTime.now();
+    }
+
+    private LocalDateTime updatedAt;
+    private boolean isDeleted = false;
+}
